Refactored to use hooks
