+++ src/components/Controls.js
@@ -1,7 +1,16 @@
 import React from 'react';
+import DeckContet from '../hooks/useDeck';
 
-const Controls = ({ dispatch }) => {
-  return <><button onClick={() => dispatch({type: 'next slide'})}>Next</button><button onClick={() => dispatch({type: 'prev slide'})}>Prev</button></>;
+const Controls = () => {
+  let deckContext = React.useContext(DeckContet);
+  return deckContext ? (
+    <>
+      <button onClick={() => dispatch({ type: 'next slide' })}>Next</button>
+      <button onClick={() => dispatch({ type: 'prev slide' })}>Prev</button>
+    </>
+  ) : (
+    <div>boop</div>
+  );
 };
 
 export default Controls;
