+++ src/components/Deck.js
@@ -1,21 +1,14 @@
 import React from 'react';
+import useDeck, { DeckContext } from '../hooks/useDeck';
 
 const initialState = { currentSlide: 0 };
 
-function reducer(state, action) {
-  switch (action.type) {
-    case 'next slide':
-      return { currentSlide: state.currentSlide + 1 };
-    case 'prev slide':
-      return { currentSlide: state.currentSlide - 1 };
-    default:
-      return { ...state };
-  }
-}
-
 const Deck = props => {
-  const [state, dispatch] = React.useReducer(reducer, initialState);
-  return props.children([dispatch, state]);
+  return (
+    <DeckContext.Provider value={useDeck(initialState)}>
+      {props.children}
+    </DeckContext.Provider>
+  );
 };
 
 export default Deck;
